plug-ins:imagemap: Use proper command title
This string is a bit tricky to find.

Go to "Menu > Filters > Web > Image Map."

This filter opens a dialog with a menu.

Go to "Menu > Tools > Use GIMP Guides..."

You can see that the command entry in the filter's menu is
"Use GIMP Guides," but it changes to
"Use Gimp Guides" in the dialog box title.

This patch fixes it.