widgets, actions: Allow multi-select in Document Dockable
This patch enables multiple selections in
the GimpDocumentView widget and
dockable. It also updates the
"documents-open" action to check if the
GimpContainerView has more than one
selection, and if so, tries to load all of
those images. The action wording is
slightly edited to indicate multiple
images can now be opened.