plug-ins: bubble the inner-format export status up.
When gimp_file_save() fails, it does not necessarily mean that there was
a plug-in error. It can just be a CANCEL status.

In particular, when we were canceling the inner-format export dialog,
GIMP was outputting an error popup saying "cancelled" as failure error
message for the file-compressor plug-in. Now it will just return with no
error, as any other cancellation.