gui, widgets: Connect ViewableDialog to style updates
This patch allows the header icon and view
in GimpViewableDialogs (and children of
this dialogue class) to scale based on user
custom icon size.
Note that it only scales when a new dialogue
is opened; it is not set to update if a dialogue
is already open and the slider is changed.
This could be added later if desired.

(cherry picked from commit aaa6bc5f9cf565d1b3db3c44907381e35dba5090)