plug-ins: crash on exporting dds with invalid number of mipmaps
Closes #15226

When you previously exported a dds with the pre-existing mipmaps,
that will be the option selected in the export dialog the next time.

If, however, that next time there were an incorrect number of mipmaps
available, the option was grayed out, but selected. Exporting then
caused a crash due to less mipmap layers being available than expected.

When writing the dds settings we need to update that setting in our
config object, so that's what we add here. Also remove updating the
local variable since that will not be used anymore in this function.

(cherry picked from commit b962bdbddd4704558ff6636f677f6f1df96c9995)