Revert "app/widgets: Guard against division by zero"
This reverts commit ee421da346e032edea78f72f1acc8924ec4fc2f4.

This is wrong. This case should not happen, or if it does, this is very
likely not the proper fix. In fact, it was trying to fix a bug created
in commit 2ba6725761 which is also wrong (and which I'm going to revert
next).

See also my comment in Gitlab:
https://gitlab.gnome.org/GNOME/gimp/-/commit/ee421da346e032edea78f72f1acc8924ec4fc2f4#note_2596612