Revert "widgets: Update tool item visibility code"
This reverts commit 2ba6725761bfa687fcab0f259712fcdd8aac0bdd.

This too is wrong. First it mentions "visible-vertical" and
"visible-horizontal" properties of GtkAction, but this is not it! Two
similarly named properties are also on GtkToolItem and these are not
deprecated! Looking at the code, tool_button is a GimpToolButton, whose
parenting indeed goes up to GtkToolItem (and that's obviously not an
action anyway).

Secondly, it creates other bugs, in particular I now sometimes have a
completely empty toolbox. That's clearly wrong.

Reverting for now. Issue #9813 will have to be fixed differently.

(cherry picked from commit 459f80a6dea2a9a65b2178781a19f3740f4aa00c)