plug-ins/filter-browser: Correctly start/stop GMainLoop
I incorrectly read the GMainLoop documentation and thought that calling
g_main_loop_new with NULL as the first parameter will return the
"default GMainLoop". This is wrong and it resulted in the original
GMainLoop to never stop and therefore for the plug-in to never stop.

This also adds a call to unref for the GMainLoop.