plug-ins: Add option to export Inkscape SVG
Inkscape SVGs have several specific properties
that affect how the SVG is shown in their
software when opened as a project file.
For now, this patch adds the option to export
as an Inkscape SVG in the GUI and saves the
layer name and group type if enabled.
More Inkscape features can be added in a
follow-up commit.