desktop: review our software description.
This is based off ongoing Teams/GIMP/CoreTeam/Discussions#9 discussion
and is meant to take care of worries which were raised in since-reverted
commit c4b4f343d4.

Basically GIMP is not a "toy software" indeed, made to just do quick
one-click edits, such as a quick crop-rotation, or apply common filters
with no options, by finger-swiping, for your social network account. It
is not to say that you cannot do all of this, but this is still part of
a much bigger feature-set, with a lot of advanced options to customize
renders, in a software made for high-end image editing and creation. The
target therefore includes graphics professionals (such as ourselves!),
but not only, which is why the reverted commit was not fine. It can be
amateurs doing high quality work too, and we don't want these creators
and usage to be left out.

The Core Team discussion is about our updated Project Vision. This is an
ongoing discussion which may span for a few more months, possibly until
the next physical meeting of contributors. So far, our Vision is
summarized as one sentence:

> GIMP is Community-driven Free Software for high-end image creation and manipulation.

This sums up several of the main aspects of GIMP: its community and how
it's being created by it, the fact it is Free Software, the fact it
targets high-end creation/manipulation, and the generic aspect of the
type of graphics works (creation or manipulation), because it is not
just about photo retouching.

This commit therefore reuses the current wording in the initial
paragraph of our metadata description. I move one example of such type
of graphics work in the second paragraph and remove the "simple" word,
as it may imply again "toy" and ourselves, as well as many other digital
painters across the world, have proved long ago that this program is not
just to do "simple painting".

This is an update before the RC release (where we will start freezing
strings). Of course, we may update further the text later (when we'll
start working on 3.4 probably), if the Project Vision further evolves.