libgimpwidgets: update color area correctly when updating color.
Since we don't necessarily re-render the color area when the color
change, let's store the actually rendered color, additionally to the
supposedly rendered color. Otherwise when changing the color by very
small increments, small enough that the color is always perceptually
identical to the previous color, we never re-render the widget, even
when the rendered color is now quite different. Indeed the "identical
color" algorithm, based on CIE2000 distance, is not transitive.

Additionally I do not check for perceptual identity in
GimpColorSelection anymore, because we also check for this in the
GimpColorArea. Better always having the right color set in the area.

This was raised in #11339 (even though the initial issue was about the
hexadecimal field).

(cherry picked from commit c558e4151788fc1ebce34a29e4516471cf3e561d)