widgets: Set ItemTreeViews insensitive with no image
Resolves #15166
Due to changes from 2.10 to 3.0, our ItemTreeViews
(Layer, Channel, Path) were not being set insensitive when
we closed all images but left the canvas open. This resulted
in CRITICALs if we try to change the opacity or the layer mode
in the LayerTreeView.
This patch adds a call to set ItemTreeViews to be insensitive if
no image is set, which also affects all the child TreeViews.