app: add options to simulate pressure and barrel rotation in MyPaint brushes.
MyPaint brushes don't have a feature similar to our brushes' dynamic and
tap directly into the pressure, tilt and barrel rotation of the input
device. This allows to e.g. paint with a mouse with a simulated
pressure. This lack was even more visible in some brushes where the
pressure was apparently changing the color too (e.g. the glow airbrush,
cf. issue #13595).

I am adding "Pressure" and "Wheel/Rotation" as a tool options for
MyPaint brush tool, because these 2 are also in MyPaint GUI (for some
reason, tilt is not available). The latter is weirdly called "Twist" in
MyPaint, but I decided to go with "Wheel/Rotation" as it's exactly the
same string we use in Dynamics (so it makes our UI more consistent hence
understandable).

Also unlike in MyPaint, I don't just add them after other brush
settings, as if part of these. I put these under a checkable frame
"Simulate Graphics Tablet Input". When checked, we use these values, and
when unchecked, we don't. Whereas in MyPaint, these are always visible,
but it depends on whether we use a device with associated feature or
not, which I find kinda confusing.