core: Create gimp_drawable_convert ()
When a drawable is converted, we check if
it has filters. If any filter has a mask, we convert it
to use the new image instead of the old one.
This prevents several crashes where the filter's mask
still pointed to the old image if it was copy/pasted
into a new one.

(cherry picked from commit 8db97f4eafd305882b0df6b1dc01cd05d553aac8)