app: fix #15382 Sensitivity ALWAYS sensitive not working as expected
A plug-in that sets Gimp.ProcedureSensitivityMask.ALWAYS is expected
to always be runnable. However, when an image is opened, but all
drawables have been removed, the plug-in is disabled.

This is because the sensitivity condition for zero drawables in an
image only set it to FALSE for the NO_DRAWABLES condition not being
met. This doesn't take into consideration that for ALWAYS, where
any image is fine, it should be set to TRUE.

To fix this we always set sensitive based on whether or not
the GIMP_PROCEDURE_SENSITIVE_NO_DRAWABLES flag is set.

(cherry picked from commit 56fa03af7a6774250ac1d4a9a5856ab3f2fd9ce6)