operations: Add translated titles to operations
The GEGL Filter Browser and API-added
filters require a "title" attribute to be defined
in the operation in order for the filter
to have a name.
The GIMP custom GEGL operations do not
have a name defined, which causes display
issues in the browser and crashes when
attempting to edit API-created filters.

This patch adds "title" attributes for four
filters which are already translated.
The remainder will need to be added after
the string freeze is over for 3.4 development.