dialogs: Canvas Size fill combo set to insensitive...
...if layers won't be resized.
This patch adds code to check if the
"Resize layers" option is set to None and
"Resize Text Layers" is turned off. If so,
then the "Fill With" combobox is not
needed and will be set to insensitive to
reduce confusion. Otherwise, the fill combo
is set to enabled.

(cherry picked from commit 897ffd5702dfe960cd20b8916b237ecee8540bd5)