Issue #14442: undo crops layer rendering out of its off-border effects.
This was happening in multiple cases, such as undoing a "Rasterize
filters" action, but also with a "Layer to Image Size" being undone, and
probably any other undo cases.

In particular here the culprit was that upon undoing, we were setting
the GimpApplicator's "gimp:compose-crop" node, while it was a "gegl:nop"
before. We must be careful not to set a crop unexpectedly when the node
was not set already, explicitly.

(cherry picked from commit 0157a9581ee01e4f757fb348bdeb3632dbbcd1b3)