Issue #14985: prevent loading recursive linked layers.
This is part of the fix for #14985, only relative to loading. It assumes
that we already have a XCF with cycling in it and will cleanly load it
(instead of forever cycling). For this to happen, I split a bit some of
the header loading code in separate functions to quickly load the header
of every link recursively until we either reach a parent file (i.e. a
cycle was discovered) or we read them all. It means we don't discover
any level of cycling (however many successive links end up to a previous
file!)

File identity will follow symbolic links as well as verify hard links.

When a cycling link is discovered, the link is downgraded to a raster
layer (showing the last render of this layer).