build: compress the .flatpak-builder/ folder in smaller parts and…
… delete build/ contents.

We were hitting an error on CI:

> ERROR: Uploading artifacts as "archive" to coordinator... 413 Request Entity Too Large  […]

Because the .flatpak-builder/ tarball was huge.

After further testing, it turned out that this is a per-file limit,
since simply not compressing the folder and passing it as artifacts
worked, except this time, we were hitting permission issues: artifacts
end up as root on successive jobs, whereas the CI job runs unprivileged
as a "build" user, which doesn't fly with flatpak-builder still needing
to work and touch previous files.

So I'm doing some in-between by compressing sub-folders (the real
advantage is not the compression part, but the archiving in tar, which
preserves ownership and permissions).

Note that while this kinda partly worked, .flatpak-builder/build/ was
still huge, even when compressing .flatpak-builder/build/* contents in
further individual archives. So I just simply drop this data as it's not
needed at all for flatpak-builder caching (cache/ subfolder is).

See also discussions in #2404.