Issue #15501: g_close failed with EBADF. This is not a valid file descriptor.
I can't reproduce this issue, because when I set a fake font file as a
broken symlink, or a file without read access, it is simply not returned
as part of my font list to start with.

But from the issue discussion and traces, this should be the right fix.

(cherry picked from commit 98bd0d16fe72d1d2b6d6cd94cb8cd77d0fbc895e)