app: no need to show -i / --no-interface in gimp-console usage.
Note that the option still exists, which allows to not error-out when
using the same set of CLI options on both the GUI and CLI binaries. We
use this ourselves in our build, which is able to use either binaries,
and if -i became forbidden, it'd make complication for our scripts.

Yet since it's basically a no-op on the CLI tool, we might as well hide
it from the usage output.