operations: Check if pixel is NaN in gimp:equalize
Resolves #3737
We use the CLAMPed value of a pixel component as part
of an operation to select an index. However, CLAMP () does
not behave as expected with NaN values, resulting in the
operation crashing when trying to use an out-of-bounds value
as an index.

This patch adds a check for isnan () to the value before using it.
If it's NaN, then we replace the value with 0.

(cherry picked from commit 97e9361ed330461a7dd8fc53728f6acbff8915a2)