gitlab-ci, build: Logs do not need to be compressed
This reverts 1fcdb724 and 8e6da721, which are inoffensive but inconsistent.
They are added while Jehan's final fix did not landed. Back then, we were
experimenting XZ and other things that we don't use in other parts of the .sh.

So, let's go back to just single tar, since logs are not that big after all.