build/windows: Move Windows .pdb utility to tools/
This reverts 63f5ea9d regarding the file naming

Since the script is an utility pretty much one-task, it suits better on tools,
similarly to what we do on babl and meson (but on these we put on root).
And, along with 8c0f92b3, it makes build/windows less scary with less files.