libgimpconfig: Fix uri -> path serialization issue
Previously, we passed a URI instead of a path into
gimp_config_path_unexpand () when serializing files.
On Windows, this causes the URI scheme to be converted
to an invalid format with too many slashes.

This patch changes the code to first attempt getting a path.
If we have one, we pass it to gimp_config_path_unexpand ()
and then convert to URI for serialization. Otherwise, we convert
just the URI without the path expansion.