build/windows: Distinguish Installer from MSIX file associations
Related to 143e4cb7, which made the MSIX on Start Menu and associations: `GIMP`
(for the stable channel).

So, this makes the Installer consistent too with just one naming on all places,
similarly to what we already do on Start Menu: `GIMP MAJOR.MINOR.MICRO-REVISION`

The naming is different from the MSIX. Otherwise, we would not know if we are
opening a file with gimp from installer or msix.

(cherry picked from commit 30680ab42e65812e829e33675ecb3ad974550301)